+++ lib/goodcheck/commands/check.rb
@@ -57,6 +57,8 @@ module Goodcheck
             end
           end
 
+          reporter.summary
+
           issue_reported ? EXIT_MATCH : EXIT_SUCCESS
         end
       end
@@ -108,7 +110,7 @@ module Goodcheck
           when DEFAULT_EXCLUSIONS.include?(path.basename.to_s)
             # noop
           when immediate || !excluded?(path)
-            path.children.each do |child|
+            path.children.sort.each do |child|
               each_file(child, &block)
             end
           end